requirements.txt
PyYAML
python-telegram-bot
telegram